refactor(cosmos): move the Cosmos ControlNet under models/controlnets
Mirrors the upstream diffusers layout (models/controlnets/controlnet_cosmos.py).
The legacy SD controlnet.py stays flat to avoid churn against dev.
