+++ src/optimum/rbln/diffusers/models/__init__.py
@@ -31,7 +31,7 @@ _import_structure = {
         "RBLNUNetSpatioTemporalConditionModel",
     ],
     "controlnet": ["RBLNControlNetModel"],
-    "controlnet_cosmos": ["RBLNCosmosControlNetModel"],
+    "controlnets": ["RBLNCosmosControlNetModel"],
     "transformers": [
         "RBLNPriorTransformer",
         "RBLNCosmosTransformer3DModel",
@@ -48,7 +48,7 @@ if TYPE_CHECKING:
         RBLNVQModel,
     )
     from .controlnet import RBLNControlNetModel
-    from .controlnet_cosmos import RBLNCosmosControlNetModel
+    from .controlnets import RBLNCosmosControlNetModel
     from .transformers import (
         RBLNCosmosTransformer3DModel,
         RBLNPriorTransformer,
