+++ src/optimum/rbln/diffusers/models/controlnets/controlnet_cosmos.py
@@ -20,16 +20,16 @@ from diffusers.models.controlnets.controlnet_cosmos import CosmosControlNetModel
 from diffusers.models.transformers.transformer_cosmos import CosmosEmbedding, CosmosPatchEmbed
 from torchvision import transforms
 
-from ...configuration_utils import RBLNCompileConfig, RBLNModelConfig
-from ...modeling import RBLNModel
-from ..configurations import RBLNCosmosControlNetModelConfig
-from .transformers.transformer_cosmos import RBLNCosmosRotaryPosEmbed, RBLNTimesteps
+from ....configuration_utils import RBLNCompileConfig, RBLNModelConfig
+from ....modeling import RBLNModel
+from ...configurations import RBLNCosmosControlNetModelConfig
+from ..transformers.transformer_cosmos import RBLNCosmosRotaryPosEmbed, RBLNTimesteps
 
 
 if TYPE_CHECKING:
     from transformers import AutoFeatureExtractor, AutoProcessor, AutoTokenizer, PretrainedConfig, PreTrainedModel
 
-    from ..modeling_diffusers import RBLNDiffusionMixin, RBLNDiffusionMixinConfig
+    from ...modeling_diffusers import RBLNDiffusionMixin, RBLNDiffusionMixinConfig
 
 
 class CosmosControlNetWrapper(torch.nn.Module):
